+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Introduced the #center method in contexts to easily find the x and y coordinates of the center point.
 - The utility method Context#each_polar_offset.
+- Palette#discretize creates a new discrete copy of a palette.
 
 ### Changed
 - Context#distance_squared now yields each distance to a block instead of writing to an array.